requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiokafka==0.10.0
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.3.0
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
asyncmy==0.2.9
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.103
botocore==1.34.103
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.4.0
crashtest==0.4.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.2
fastjsonschema==2.19.1
filelock==3.14.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
git-filter-repo==2.38.0
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==7.1.0
installer==0.7.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kafka-python==2.0.2
keyring==24.3.1
Levenshtein==0.25.1
mangum==0.17.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
more-itertools==10.2.0
motor==3.4.0
msgpack==1.0.8
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pipx==1.5.0
pkginfo==1.10.0
platformdirs==4.2.1
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.2
PyMySQL==1.1.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
tenacity==8.3.0
tomlkit==0.12.5
tornado==6.4
traitlets==5.14.3
trove-classifiers==2024.5.22
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
userpath==1.9.2
uvicorn==0.29.0
virtualenv==20.26.2
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.18.2
